+++ Editor/Sources/RecorderAnalytics.cs
@@ -4,7 +4,7 @@ using System.Runtime.CompilerServices;
 using System;
 using System.Collections.Generic;
 using System.Text;
-using Unity.Media;
+using UnityEditor.Recorder.AOV;
 using UnityEditor.Recorder.Encoder;
 using UnityEditor.Recorder.Input;
 using UnityEngine;
@@ -25,6 +25,9 @@ namespace UnityEditor.Recorder
         const string vendorKey = "unity.recorder";
         const string startEventName = "recorder_session_start";
         const string completeEventName = "recorder_session_complete";
+
+        const int startEventVersion = 2;
+        const int completeEventVersion = 1;
         [Serializable]
         internal struct RecorderSessionStartEvent
         {
@@ -41,7 +44,8 @@ namespace UnityEditor.Recorder
             public List<AnimationRecorderInfo> animation_recorder_info;
             public List<ImageRecorderInfo> image_recorder_info;
             public List<MovieRecorderInfo> movie_recorder_info;
-            public List<GifRecorderInfo> gif_recorder_info;
+            // need to keep as to not mess up the schema
+            public List<AOVImageRecorderInfo> aov_recorder_info;
         }
 
         [Serializable]
@@ -62,106 +66,41 @@ namespace UnityEditor.Recorder
 
 
         [Serializable]
-        internal struct RecorderInfo
+        internal class RecorderInfo
         {
             public string type;
             public string record_guid;
+            public bool enabled;
 
-            public static RecorderInfo FromRecorder(Recorder r)
+            public static void FromRecorder(RecorderSettings r, RecorderInfo recorderInfo)
             {
-                return new RecorderInfo
-                {
-                    type = r.GetType().FullName,
-                    record_guid = r.GetInstanceID().ToString(),
-                };
+                var errors = new List<string>();
+                r.GetErrors(errors);
+
+                recorderInfo.type = RecordersInventory.GetRecorderInfo(r.GetType()).recorderType.FullName;
+                recorderInfo.record_guid = r.GetInstanceID().ToString();
+                recorderInfo.enabled = r.Enabled;
             }
         }
 
         [Serializable]
-        internal struct AnimationRecorderInfo
+        internal class AnimationRecorderInfo : RecorderInfo
         {
-            public string type;
-            public string record_guid;
             public bool record_hierarchy;
             public bool  clamped_tangets;
             public string anim_compression;
-            public static AnimationRecorderInfo FromRecorder(AnimationRecorder r)
+            public static AnimationRecorderInfo FromRecorder(AnimationRecorderSettings r)
             {
-                return new AnimationRecorderInfo
+                var ret = new AnimationRecorderInfo
                 {
-                    type = r.GetType().FullName,
+                    type = RecordersInventory.GetRecorderInfo(r.GetType()).recorderType.FullName,
                     record_guid = r.GetInstanceID().ToString(),
-                    anim_compression =  r.Settings.AnimationInputSettings.SimplyCurves.ConvertToString(),
-                    record_hierarchy = r.Settings.AnimationInputSettings.Recursive,
-                    clamped_tangets = r.Settings.AnimationInputSettings.ClampedTangents
-                };
-            }
-        }
-
-        [Serializable]
-        internal struct ImageRecorderInfo
-        {
-            public string type;
-            public string record_guid;
-            public string source;
-            public int output_resolution_w;
-            public int output_resolution_h;
-            public string media_format;
-            public string color_space;
-            public static ImageRecorderInfo FromRecorder(ImageRecorder r)
-            {
-                return new ImageRecorderInfo()
-                {
-                    type = r.GetType().FullName,
-                    record_guid = r.GetInstanceID().ToString(),
-                    color_space = r.Settings.m_ColorSpace.ConvertToString(),
-                    media_format = r.Settings.OutputFormat.ConvertToString(),
-                    output_resolution_h = r.Settings.imageInputSettings.OutputHeight,
-                    output_resolution_w = r.Settings.imageInputSettings.OutputWidth,
-                    source = r.Settings.imageInputSettings.ConvertToString()
-                };
-            }
-        }
-
-        [Serializable]
-        internal struct MovieRecorderInfo
-        {
-            public string type;
-            public string record_guid;
-            public string source;
-            public int output_resolution_w;
-            public int output_resolution_h;
-            public string media_format;
-            public bool include_audio;
-            public string codec_format;
-            public string quality;
-            public static MovieRecorderInfo FromRecorder(MovieRecorder r)
-            {
-                var ret = new MovieRecorderInfo()
-                {
-                    type = r.GetType().FullName,
-                    record_guid = r.GetInstanceID().ToString(),
-                    output_resolution_h = r.Settings.ImageInputSettings.OutputHeight,
-                    output_resolution_w = r.Settings.ImageInputSettings.OutputWidth,
-                    source = r.Settings.ImageInputSettings.ConvertToString(),
-                    include_audio = r.Settings.CaptureAudio,
+                    anim_compression =  r.AnimationInputSettings.SimplyCurves.ConvertToString(),
+                    record_hierarchy = r.AnimationInputSettings.Recursive,
+                    clamped_tangets = r.AnimationInputSettings.ClampedTangents
                 };
 
-                if (r.Settings.EncoderSettings != null)
-                {
-                    if (r.Settings.EncoderSettings.GetType() == typeof(ProResEncoderSettings))
-                    {
-                        ret.codec_format =
-                            ((ProResEncoderSettings)(r.Settings.EncoderSettings)).Format.ConvertToString();
-                        ret.media_format = "MOV";
-                    }
-                    else if (r.Settings.EncoderSettings.GetType() == typeof(CoreEncoderSettings))
-                    {
-                        ret.media_format = ((CoreEncoderSettings)(r.Settings.EncoderSettings)).Codec.ConvertToString();
-                        ret.quality = ((CoreEncoderSettings)(r.Settings.EncoderSettings)).EncodingQuality
-                            .ConvertToString();
-                    }
-                }
+                RecorderInfo.FromRecorder(r, ret);
 
                 return ret;
             }
@@ -169,23 +108,162 @@ namespace UnityEditor.Recorder
 
 
         [Serializable]
-        internal struct GifRecorderInfo
+        internal class ImageRecorderInfo : RecorderInfo
         {
-            public string type;
-            public string record_guid;
             public string source;
             public int output_resolution_w;
             public int output_resolution_h;
+            public string media_format;
+            public string color_space;
+            public AccumulationInfo accumulation_info;
 
-            public static GifRecorderInfo FromRecorder(GIFRecorder r)
+            public static ImageRecorderInfo FromRecorder(ImageRecorderSettings r)
             {
-                return new GifRecorderInfo()
+                var ret = new ImageRecorderInfo()
                 {
-                    type = r.GetType().FullName,
+                    type = RecordersInventory.GetRecorderInfo(r.GetType()).recorderType.FullName,
                     record_guid = r.GetInstanceID().ToString(),
-                    output_resolution_h = r.Settings.imageInputSettings.OutputHeight,
-                    output_resolution_w = r.Settings.imageInputSettings.OutputWidth,
-                    source = r.Settings.imageInputSettings.ConvertToString()
+                    color_space = r.OutputFormat == ImageRecorderSettings.ImageRecorderOutputFormat.EXR ? r.OutputColorSpace.ConvertToString() : null,
+                    media_format = r.OutputFormat.ConvertToString(),
+                    output_resolution_h = r.imageInputSettings.OutputHeight,
+                    output_resolution_w = r.imageInputSettings.OutputWidth,
+                    source = r.imageInputSettings.ConvertToString(),
+                    accumulation_info = AccumulationInfo.FromRecorder(r)
+                };
+                RecorderInfo.FromRecorder(r, ret);
+
+                return ret;
+            }
+        }
+
+        [Serializable]
+        internal class MovieRecorderInfo : RecorderInfo
+        {
+            public string source;
+            public int output_resolution_w;
+            public int output_resolution_h;
+            public string media_format;
+            public bool include_audio;
+            public string codec_format;
+            public string quality;
+            public AccumulationInfo accumulation_info;
+            public string encoder_type;
+            public float target_bitrate;
+            public uint keyframe_distance;
+            public string encoding_profile;
+            public uint b_frames;
+
+            public static MovieRecorderInfo FromRecorder(MovieRecorderSettings r)
+            {
+                var ret = new MovieRecorderInfo()
+                {
+                    type = RecordersInventory.GetRecorderInfo(r.GetType()).recorderType.FullName,
+                    record_guid = r.GetInstanceID().ToString(),
+                    output_resolution_h = r.ImageInputSettings.OutputHeight,
+                    output_resolution_w = r.ImageInputSettings.OutputWidth,
+                    source = r.ImageInputSettings.ConvertToString(),
+                    include_audio = r.CaptureAudio,
+                    accumulation_info = AccumulationInfo.FromRecorder(r),
+                    encoder_type = r.EncoderSettings != null
+                        ? r.EncoderSettings.GetType().FullName
+                        : ""
+                };
+
+                if (r.EncoderSettings == null) return ret;
+
+                switch (r.EncoderSettings)
+                {
+                    case ProResEncoderSettings proResEncoderSettings:
+                        ret.codec_format =
+                            proResEncoderSettings.Format.ConvertToString();
+
+                        break;
+                    case CoreEncoderSettings coreEncoderSettings:
+                    {
+                        ret.media_format = coreEncoderSettings.Codec.ConvertToString();
+                        ret.quality = coreEncoderSettings.EncodingQuality
+                            .ConvertToString();
+                        if (coreEncoderSettings.EncodingQuality == CoreEncoderSettings.VideoEncodingQuality.Custom)
+                        {
+                            ret.target_bitrate = coreEncoderSettings.TargetBitRate;
+                            ret.keyframe_distance = coreEncoderSettings.Codec == CoreEncoderSettings.OutputCodec.WEBM
+                                ? coreEncoderSettings.KeyframeDistance
+                                : coreEncoderSettings.GopSize;
+                            if (coreEncoderSettings.Codec == CoreEncoderSettings.OutputCodec.MP4)
+                            {
+                                ret.encoding_profile = coreEncoderSettings.EncodingProfile.ConvertToString();
+                                ret.b_frames = coreEncoderSettings.numConsecutiveBFrames;
+                            }
+                        }
+
+                        break;
+                    }
+                }
+
+                ret.media_format = r.EncoderSettings.Extension;
+
+                RecorderInfo.FromRecorder(r, ret);
+                return ret;
+            }
+        }
+
+        [Serializable]
+        internal class AOVImageRecorderInfo : RecorderInfo
+        {
+            public int output_resolution_w;
+            public int output_resolution_h;
+            public string media_format;
+            public string color_space;
+            public string aov;
+
+            public static AOVImageRecorderInfo FromRecorder(AOVRecorderSettings r)
+            {
+                var ret = new AOVImageRecorderInfo()
+                {
+                    type = RecordersInventory.GetRecorderInfo(r.GetType()).recorderType.FullName,
+                    record_guid = r.GetInstanceID().ToString(),
+                    color_space = r.OutputFormat == ImageRecorderSettings.ImageRecorderOutputFormat.EXR ? r.OutputColorSpace.ConvertToString() : null,
+                    media_format = r.OutputFormat.ConvertToString(),
+                    output_resolution_h = r.imageInputSettings.OutputHeight,
+                    output_resolution_w = r.imageInputSettings.OutputWidth,
+                    aov = r.AOVSelection.ToString()
+                };
+
+                RecorderInfo.FromRecorder(r, ret);
+
+                return ret;
+            }
+        }
+
+        [Serializable]
+        internal struct AccumulationInfo
+        {
+            public bool enabled;
+            public int num_samples;
+            public string shutter_profile_range;
+            public float shutter_begin;
+            public float shutter_end;
+            public bool subpixel_jitter;
+            public float shutter_interval;
+
+            public static AccumulationInfo FromRecorder(RecorderSettings r)
+            {
+                if (r is not IAccumulation settings)
+                    return default;
+                var aSettings = settings.GetAccumulationSettings();
+                return new AccumulationInfo
+                {
+                    enabled = aSettings.CaptureAccumulation,
+                    num_samples = aSettings.Samples,
+                    subpixel_jitter = aSettings.UseSubPixelJitter,
+                    shutter_profile_range = aSettings.ShutterType.ToString(),
+                    shutter_interval = aSettings.ShutterInterval,
+                    shutter_begin = aSettings.ShutterType == AccumulationSettings.ShutterProfileType.Range
+                        ? aSettings.ShutterFullyOpen
+                        : 0,
+                    shutter_end = aSettings.ShutterType == AccumulationSettings.ShutterProfileType.Range
+                        ? aSettings.ShutterBeginsClosing
+                        : 0
                 };
             }
         }
@@ -196,10 +274,10 @@ namespace UnityEditor.Recorder
             if (!EditorAnalytics.enabled)
                 return;
 
-            EditorAnalytics.RegisterEventWithLimit(startEventName, maxEventsPerHour, maxNumberOfElements, vendorKey);
+            EditorAnalytics.RegisterEventWithLimit(startEventName, maxEventsPerHour, maxNumberOfElements, vendorKey, startEventVersion);
 
             var data = CreateSessionStartEvent(controller);
-            EditorAnalytics.SendEventWithLimit(startEventName, data);
+            EditorAnalytics.SendEventWithLimit(startEventName, data, startEventVersion);
         }
 
         public static void SendStopEvent(RecorderController controller, bool error)
@@ -207,10 +285,10 @@ namespace UnityEditor.Recorder
             if (!EditorAnalytics.enabled)
                 return;
 
-            EditorAnalytics.RegisterEventWithLimit(completeEventName, maxEventsPerHour, maxNumberOfElements, vendorKey);
+            EditorAnalytics.RegisterEventWithLimit(completeEventName, maxEventsPerHour, maxNumberOfElements, vendorKey, completeEventVersion);
 
             var data = CreateStopEvent(controller, error);
-            EditorAnalytics.SendEventWithLimit(completeEventName, data);
+            EditorAnalytics.SendEventWithLimit(completeEventName, data, completeEventVersion);
         }
 
         // Used by the Timeline
@@ -219,10 +297,10 @@ namespace UnityEditor.Recorder
             if (!EditorAnalytics.enabled)
                 return;
 
-            EditorAnalytics.RegisterEventWithLimit(startEventName, maxEventsPerHour, maxNumberOfElements, vendorKey);
+            EditorAnalytics.RegisterEventWithLimit(startEventName, maxEventsPerHour, maxNumberOfElements, vendorKey, startEventVersion);
             var data = CreateSessionStartEvent(session);
             // Send the data to the database
-            EditorAnalytics.SendEventWithLimit(startEventName, data);
+            EditorAnalytics.SendEventWithLimit(startEventName, data, startEventVersion);
         }
 
         public static void SendStopEvent(RecordingSession session, bool error, bool complete)
@@ -230,12 +308,12 @@ namespace UnityEditor.Recorder
             if (!EditorAnalytics.enabled)
                 return;
 
-            EditorAnalytics.RegisterEventWithLimit(completeEventName, maxEventsPerHour, maxNumberOfElements, vendorKey);
+            EditorAnalytics.RegisterEventWithLimit(completeEventName, maxEventsPerHour, maxNumberOfElements, vendorKey, completeEventVersion);
 
             var data = CreateStopEvent(session, error, complete);
 
             // Send the data to the database
-            EditorAnalytics.SendEventWithLimit(completeEventName, data);
+            EditorAnalytics.SendEventWithLimit(completeEventName, data, completeEventVersion);
         }
 
         internal static RecorderSessionEndEvent CreateStopEvent(RecorderController controller, bool error)
@@ -330,12 +408,12 @@ namespace UnityEditor.Recorder
             };
 
             GetSpecificRecorderInfos(
-                new[] {session.recorder},
+                new[] {session.recorder.settings},
                 out data.animation_recorder_info,
                 out data.image_recorder_info,
                 out data.movie_recorder_info,
-                out data.gif_recorder_info,
-                out data.recorder_info);
+                out data.recorder_info,
+                out data.aov_recorder_info);
 #if DEBUG_ANALYTICS
             var json = JsonUtility.ToJson(data, prettyPrint: true);
             Debug.Log(json);
@@ -364,12 +442,12 @@ namespace UnityEditor.Recorder
             };
 
             GetSpecificRecorderInfos(
-                controller.m_RecordingSessions.Select(x => x.recorder).Where(x => x.settings.Enabled),
+                controller.Settings.RecorderSettings,
                 out data.animation_recorder_info,
                 out data.image_recorder_info,
                 out data.movie_recorder_info,
-                out data.gif_recorder_info,
-                out data.recorder_info);
+                out data.recorder_info,
+                out data.aov_recorder_info);
 
 #if DEBUG_ANALYTICS
             var json = JsonUtility.ToJson(data, prettyPrint: true);
@@ -378,33 +456,35 @@ namespace UnityEditor.Recorder
             return data;
         }
 
-        static void GetSpecificRecorderInfos(IEnumerable<Recorder> recorders, out List<AnimationRecorderInfo> anim,
-            out List<ImageRecorderInfo> image, out List<MovieRecorderInfo> movie, out List<GifRecorderInfo> gif, out List<RecorderInfo> recorder)
+        static void GetSpecificRecorderInfos(IEnumerable<RecorderSettings> recorders, out List<AnimationRecorderInfo> anim,
+            out List<ImageRecorderInfo> image, out List<MovieRecorderInfo> movie, out List<RecorderInfo> recorder, out List<AOVImageRecorderInfo> aov)
         {
             anim = new List<AnimationRecorderInfo>();
             image = new List<ImageRecorderInfo>();
             movie = new List<MovieRecorderInfo>();
-            gif = new List<GifRecorderInfo>();
             recorder = new List<RecorderInfo>();
+            aov = new List<AOVImageRecorderInfo>();
 
             foreach (var reco in recorders)
             {
                 switch (reco)
                 {
-                    case AnimationRecorder r:
+                    case AnimationRecorderSettings r:
                         anim.Add(AnimationRecorderInfo.FromRecorder(r));
                         break;
-                    case ImageRecorder r:
+                    case ImageRecorderSettings r:
                         image.Add(ImageRecorderInfo.FromRecorder(r));
                         break;
-                    case MovieRecorder r:
+                    case MovieRecorderSettings r:
                         movie.Add(MovieRecorderInfo.FromRecorder(r));
                         break;
-                    case GIFRecorder r:
-                        gif.Add(GifRecorderInfo.FromRecorder(r));
+                    case AOVRecorderSettings r:
+                        aov.Add(AOVImageRecorderInfo.FromRecorder(r));
                         break;
                     default:
-                        recorder.Add(RecorderInfo.FromRecorder(reco));
+                        var ri = new RecorderInfo();
+                        RecorderInfo.FromRecorder(reco, ri);
+                        recorder.Add(ri);
                         break;
                 }
             }
@@ -415,10 +495,10 @@ namespace UnityEditor.Recorder
                 image = null;
             if (movie.Count == 0)
                 movie = null;
-            if (gif.Count == 0)
-                gif = null;
             if (recorder.Count == 0)
                 recorder = null;
+            if (aov.Count == 0)
+                aov = null;
         }
 
         internal static string ConvertToString<T>(this T e) where T : Enum